+++ tests/testthat/test-ibr_index.R
@@ -1,5 +1,6 @@
 test_that("assess_ibr_index", {
-  expected <- ibr_index(enzact, enzact_coef)
+  outstd <- ibr_std(enzact)
+  expected <- ibr_index(outstd, enzact_coef)
   expect_type(expected, "list")
   expect_silent(enzact)
 })